pyproject.toml
[project]
name = "document-parsing"
version = "0.1.0"
description = ""
authors = [
    {name = "Malay Raj Bareth",email = "malay@arkvedaai.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch (==2.10.0+cpu)",
    "torchvision (==0.25.0+cpu)",
    "torchaudio (==2.10.0+cpu)",
    "accelerate (==1.12.0)",
    "einops (==0.8.2)",
    "h5py (==3.15.1)",
    "numpy (==1.26.4)",
    "onnxruntime (==1.20.1)",
    "opencv-python (==4.11.0.86)",
    "pdf2image (==1.17.0)",
    "pillow (==10.4.0)",
    "pydantic (==2.12.5)",
    "pydantic-settings (==2.12.0)",
    "pymupdf (==1.26.7)",
    "pypdfium2 (==4.30.0)",
    "python-dotenv (==1.2.1)",
    "pyyaml (==6.0.3)",
    "rapid-latex-ocr (==0.0.9)",
    "rapidocr-onnxruntime (==1.4.4)",
    "shapely (==2.1.2)",
    "super-image (==0.2.0)",
    "surya-ocr (==0.17.1)",
    "sympy (==1.14.0)",
    "transformers (==4.57.6)",
    "easyocr (>=1.7.2,<2.0.0)",
    "ruff (>=0.15.2,<0.16.0)",
    "bandit (>=1.9.4,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}
torchaudio = {source = "pytorch"}
